Add 'undo' and 'redo' translations.
Undo and redo were missing from the translation so I added them. I also modified some of the other translations to be more in line with what is generally used in finnish. For example "Submit" means the same as "Send" in finnish so it was translated as such, even though a more context-specific translation is "Lisää" which means "Add".