Fix Cleanpaste plugin window freeze #618
Create minLength var that stores the shortest length of htmlBefore and htmlAfter, and ends comparison at that length if no differences are discovered by then. charAt will return an empty string if addressing beyond the end of a string, which can create an endless loop in this case.

Fix adds implicit dependency on javascript Math lib, and creates two lines that exceed 80 characters in width.