openModalInsert:
Fix bug where, if a name is provided which is different to the fieldName, an exception was thrown in the OK button handler.

When populating the results in values, use name (if provided) rather than the fieldName.

Explicitly support radio buttons (multiple type:'radio' options with the same name), returning the name element in values to the value of the selected radio button.

To compensate a little for the few extra lines of code, optimise the $.each which builds the dialog html slightly.

Add a little more documentation in docs/documentation/core/index/html and docs/documentation/index.html (I was not sure whether one was generated automatically from the other, so updated both).

*** NOTE ***

I have not updated any of the files in dist. I assume that is done as part of the release process.