used the style attribute to make the resized images is more responsive, sync trumbowyg code on the canvas tag blur
When the tag canvas is replaced by the tag img, the style attribute of the img tag is set with css code that allow it to be more responsive for mobile devices.

Added the blur event on the tag canvas to avoid issue when a user keep the resizeimg plugin active over an image: this cause the saving of the canvas tag in the HTML code of trumbowyg and the absence of the image on the next code reload, sobstituted by an empty square.