Document adding custom icons for standalone plugins.
The previously recommended way of doing this was to roll your own build of
Trumbowyg, but for lazy people like me, that looks too much like effort.

It also implies that plugins *cannot* be distributed standalone (e.g. load
Trumbowyg from a CDN, then your own plugin on top of that). This is not so!
But that notion might put someone off using Trumbowyg for their project.
Trumbowyg doesn't actually care where the button's symbol comes from; if
it exists in the document with the correct ID it will work.